anno_doumi/agent/requirements.txt
numpy
langchain
langchain-community
openai
openai-whisper
webrtcvad
sounddevice
python-dotenv
